Feature/search product (#5)
* Authenticate and Authorization using spring security and JWT token

* Create new Product Detail API

* Create new Product Detail API

* Create new Product Detail API

* Create new Product Image Upload part added

* Create new Product Image Upload part added

* product save api add security

* get All product API implement without security

* Delete product details API implement for hard delete

* Get Product Details using Filter ProductId implement API

* Admin Product API'S Security enable completed and finished product service Feature.

* get Product all API's disabled security

* product entity table change description length

* product entity table change description length

* Place Order API implement

* Place Order API implement

* User access get Product detail API implement

* User access get Product detail API implement

* Sonar issue fixed

* Home Pagination implemented

* Search Product API implement using hibernate

+++ src/main/java/com/sliat/crm/ecommerce/service/impl/ProductServiceImpl.java
@@ -5,20 +5,29 @@ import com.sliat.crm.ecommerce.dao.ProductDao;
 import com.sliat.crm.ecommerce.dto.ProductDto;
 import com.sliat.crm.ecommerce.entity.Product;
 import com.sliat.crm.ecommerce.service.ProductService;
-import org.springframework.beans.factory.annotation.Autowired;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 
+
+
 @Service
+@RequiredArgsConstructor
 public class ProductServiceImpl implements ProductService {
 
-    @Autowired
-    private ObjectMapper mapper;
 
-    @Autowired
-    private ProductDao productDao;
+    private final ObjectMapper mapper;
+
+
+    private final ProductDao productDao;
+
+
 
     @Override
     public ProductDto createNewProduct(ProductDto productData) {
@@ -28,19 +37,42 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public List<ProductDto> getAllProduct() {
+
+    public List<ProductDto> getAllProduct(Integer pageNumber, String searchKey) {
         List<ProductDto> productDtoList = new ArrayList<>();
-        productDao.findAll().forEach(product -> {
-            if (product instanceof Product)
-                productDtoList.add(mapper.convertValue(product, ProductDto.class));
-        });
+        Pageable pageRequest = PageRequest.of(pageNumber, 10);
+        if (searchKey.equals("")) {
+            productDao.findAll(pageRequest).forEach(product -> {
+                if (product instanceof Product)
+                    productDtoList.add(mapper.convertValue(product, ProductDto.class));
+            });
+        } else {
+
+            List<Product> products = productDao.findByNameContainingIgnoreCaseOrDescriptionContainingIgnoreCase(searchKey, searchKey, pageRequest);
+            if (products == null)
+                return productDtoList;
+            products.forEach(product ->
+                    productDtoList.add(mapper.convertValue(product, ProductDto.class))
+            );
+
+        }
 
         return productDtoList;
     }
 
     @Override
-    public void deleteProductDetails(Integer id) {
-        productDao.deleteById(id);
+
+    public boolean deleteProductDetails(Integer id) {
+        if (productDao.findById(id).isPresent()) {
+            productDao.deleteById(id);
+            return true;
+        }
+
+        return false;
+
+
+
+
     }
 
     @Override
@@ -51,4 +83,23 @@ public class ProductServiceImpl implements ProductService {
 
         return null;
     }
+
+
+    @Override
+    public List<ProductDto> getProductDetail(boolean isSingleProductDetail, Integer productId) {
+        if (isSingleProductDetail) {
+            List<ProductDto> list = new ArrayList<>();
+            Product product = productDao.findById(productId).orElse(null);
+            if (product != null)
+                list.add(mapper.convertValue(product, ProductDto.class));
+
+            return list;
+        } else {
+            //check out entire cart
+            //TODO
+        }
+
+        return new ArrayList<>();
+    }
+
 }
